server: do not stick to IP-based security for zone transfers
Since the zone transfers are secured by TSIG, accept notify from any IP address
as long as they are signed. This increases the risk a secondary may send a
signed SOA request to the primary, even if the notify was sent by a third party.
A signed SOA isn't that bad, and indeed only the secondary <-> primary link can
be targeted (the secondary won't ask random IP addresses for this).

This should allow a secondary to run behind a proxy.